Added new commands & operators
Added move, target and useItem commands.
Made selectors, operators and commands case-insensitive
Added rnd and rndI operators

+++ Common/src/net/minecraft/command/collections/TypeIDs.java
@@ -237,7 +237,7 @@ public final class TypeIDs
 			@Override
 			public Boolean convert(final Integer toConvert) throws CommandException
 			{
-				return toConvert != 0;
+				return toConvert.intValue() != 0;
 			}
 		});
